#1185 feat: move testEmoji to ObjectsTableTest + minor fixes
